+++ samplesheets/models.py
@@ -997,8 +997,10 @@ class Process(NodeMixin, BaseSampleSheet):
 
 
 class ISATab(models.Model):
-    """Class for storing ISA-Tab files for one investigation, including its
-    studies and assays"""
+    """
+    Class for storing ISA-Tab files for one investigation, including its
+    studies and assays.
+    """
 
     #: Project to which the ISA-Tab belongs
     project = models.ForeignKey(
@@ -1048,9 +1050,9 @@ class ISATab(models.Model):
         on_delete=models.CASCADE,
     )
 
-    #: DateTime of ISA-Tab creation
+    #: DateTime of ISA-Tab creation or restoring
     date_created = models.DateTimeField(
-        auto_now=True, help_text='DateTime of ISAtab creation'
+        auto_now_add=True, help_text='DateTime of ISAtab creation or restoring'
     )
 
     #: Version of altamISA used when processing this ISA-Tab
@@ -1063,6 +1065,14 @@ class ISATab(models.Model):
 
     #: Optional extra data
     extra_data = models.JSONField(default=dict, help_text='Optional extra data')
+
+    #: Short Description (optional)
+    description = models.CharField(
+        max_length=128,
+        blank=True,
+        null=True,
+        help_text='Short description for ISA-Tab version (optional)',
+    )
 
     #: Internal UUID for the object
     sodar_uuid = models.UUIDField(
@@ -1081,6 +1091,11 @@ class ISATab(models.Model):
     # Custom row-level functions
 
     def get_name(self):
+        """Return version timestamp as its name"""
+        return localtime(self.date_created).strftime('%Y-%m-%d %H:%M:%S')
+
+    def get_full_name(self):
+        """Return full name with investigation title or archive name"""
         investigation = Investigation.objects.filter(
             sodar_uuid=self.investigation_uuid
         ).first()
@@ -1090,9 +1105,7 @@ class ISATab(models.Model):
             name = self.archive_name.split('.')[0]
         else:
             name = self.project.title
-        return name + ' ({})'.format(
-            localtime(self.date_created).strftime('%Y-%m-%d %H:%M:%S')
-        )
+        return name + ' ({})'.format(self.get_name())
 
 
 class IrodsAccessTicketActiveManager(models.Manager):