+++ samplesheets/templatetags/samplesheets_tags.py
@@ -123,12 +123,9 @@ def get_isatab_tag_html(isatab):
         return '<span class="text-muted">N/A</span>'
     ret = ''
     for tag in sorted(isatab.tags):
-        ret += (
-            '<span class="badge badge-pill badge-{} mr-1">'
-            '{}</span>\n'.format(
-                TAG_COLORS[tag] if tag in TAG_COLORS else DEFAULT_TAG_COLOR,
-                tag.capitalize(),
-            )
+        ret += '<span class="badge badge-pill badge-{}">' '{}</span>\n'.format(
+            TAG_COLORS[tag] if tag in TAG_COLORS else DEFAULT_TAG_COLOR,
+            tag.capitalize(),
         )
     return ret